+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [6.2.2] - 2023-04-26
+### Changed
+- The list for selecting codes in the "pull data for cohort" step is limited to
+  those codes that exist for the patient cohort.
+
 ## [6.2.1] - 2023-04-25
 ### Changed
 - Filter certain columns by a prefix in dbGaP Variables table.